Make sortFile a separately function + add cli and gui modes + make progress bar trying to work
